fix(proto): nest fks-proto modules + align distributed schema to call sites
1. prost generated 'super::super::common::v1::Side' for cross-package field refs in fks.execution.v1. src/lib.rs flat-sibling modules made that overshoot the crate root. Fix: nest modules under crate::fks::* matching each proto package path, then re-export the flat names consumers already import.

2. distributed.proto NodeCapabilities / ResourceUsage were vendored with field names from a different upstream version (cpu_percent / memory_bytes / gpu_models) while grpc_client.rs constructs them with has_cuda / gpu_type / total_memory / cpu_usage / gpu_usage(Vec) / network_bandwidth. Retagged the proto to match call sites.

3. Three u64 -> i64 casts in regime_bridge_{consumer,server} where AtomicU64 fetch_add output reached the i64 sequence field.

4. crates/lob/Cargo.toml: criterion was workspace = true but never declared in workspace.dependencies. Switched to pinned 0.5.

Adds .gitignore so target/ doesn't pollute future commits.

cargo check --workspace clean; 539 janus-forward + 116 fks-proto/janus-core unit tests pass.

+++ services/forward/src/regime_bridge_consumer.rs
@@ -386,7 +386,7 @@ impl ServerHandler {
 
 impl BridgeStateHandler for ServerHandler {
     async fn handle(&self, stream_id: &str, state: &StreamedBridgeState) -> Result<()> {
-        let seq = self.sequence.fetch_add(1, Ordering::Relaxed);
+        let seq = self.sequence.fetch_add(1, Ordering::Relaxed) as i64;
         let ingested = self.total_ingested.fetch_add(1, Ordering::Relaxed) + 1;
 
         // Build a proto RegimeState and push it through the embedded server
@@ -551,7 +551,7 @@ impl GrpcForwarderHandler {
     fn to_proto_state(&self, state: &StreamedBridgeState) -> RegimeState {
         let hypo = parse_hypothalamus_regime(&state.hypothalamus_regime);
         let amyg = parse_amygdala_regime(&state.amygdala_regime);
-        let seq = self.sequence.fetch_add(1, Ordering::Relaxed);
+        let seq = self.sequence.fetch_add(1, Ordering::Relaxed) as i64;
 
         RegimeState {
             symbol: state.symbol.clone(),
